+++ src/proposal.rs
@@ -30,7 +30,7 @@ pub fn prepare_proposal(helper: &UpgradeHelper) -> Result<String, RenderError> {
         "network": format!("{}", helper.network), // TODO: implement serialize trait here?
         "previous_version": get_release_md_link(helper.previous_version.as_str()),
         "version": get_release_md_link(helper.target_version.as_str()),
-        "voting_time": helper.voting_period.num_hours(),
+        "voting_time": helper.voting_period,
     });
 
     handlebars.render("proposal", &data)
@@ -39,10 +39,8 @@ pub fn prepare_proposal(helper: &UpgradeHelper) -> Result<String, RenderError> {
 /// Writes the proposal contents to a file.
 pub fn write_proposal_to_file(
     proposal: &str,
-    network: Network,
-    target_version: &str,
+    proposal_file_name: &str,
 ) -> Result<(), std::io::Error> {
-    let proposal_file_name = format!("proposal-{}-{}.md", network, target_version);
     std::fs::write(proposal_file_name, proposal)
 }
 
@@ -50,15 +48,27 @@ pub fn write_proposal_to_file(
 fn get_height_with_link(network: Network, height: u64) -> String {
     let height_with_commas = height.to_formatted_string(&num_format::Locale::en);
     match network {
-        Network::LocalNode => format!("[{}](https://www.mintscan.io/evmos/blocks/{})", height_with_commas, height),
-        Network::Mainnet => format!("[{}](https://www.mintscan.io/evmos/blocks/{})", height_with_commas, height),
-        Network::Testnet => format!("[{}](https://testnet.mintscan.io/evmos-testnet/blocks/{})", height_with_commas, height),
+        Network::LocalNode => format!(
+            "[{}](https://www.mintscan.io/evmos/blocks/{})",
+            height_with_commas, height
+        ),
+        Network::Mainnet => format!(
+            "[{}](https://www.mintscan.io/evmos/blocks/{})",
+            height_with_commas, height
+        ),
+        Network::Testnet => format!(
+            "[{}](https://testnet.mintscan.io/evmos-testnet/blocks/{})",
+            height_with_commas, height
+        ),
     }
 }
 
 /// Returns the appropriate Markdown link to the release on GitHub for the given version.
 fn get_release_md_link(version: &str) -> String {
-    format!("[{0}](https://github.com/evmos/evmos/releases/tag/{0})", version)
+    format!(
+        "[{0}](https://github.com/evmos/evmos/releases/tag/{0})",
+        version
+    )
 }
 
 #[cfg(test)]
@@ -80,7 +90,8 @@ mod tests {
 
     #[test]
     fn test_write_proposal_to_file_pass() {
-        let result = write_proposal_to_file("test", Network::Mainnet, "v0.1.0");
+        let proposal_file_name = format!("proposal-{}-{}.md", Network::Mainnet, "v0.1.0");
+        let result = write_proposal_to_file("test", proposal_file_name.as_str());
         assert!(
             result.is_ok(),
             "Error writing proposal to file: {}",
@@ -88,7 +99,6 @@ mod tests {
         );
 
         // Check that file exists
-        let proposal_file_name = format!("proposal-{}-{}.md", Network::Mainnet, "v0.1.0");
         assert!(
             std::path::Path::new(proposal_file_name.as_str()).exists(),
             "Proposal file does not exist",