+++ CHANGELOG.md
@@ -2,12 +2,16 @@
 
 ## Unreleased
 
+### Features
+
+- [#16](https://github.com/MalteHerrmann/upgrade-helper/pull/16) Prepare CLI command to submit proposal.
+- [#5](https://github.com/MalteHerrmann/upgrade-helper/pull/5) Prepare proposal contents depending on input.
+- [#3](https://github.com/MalteHerrmann/upgrade-helper/pull/3) Add basic CLI structure.
+
 ### Improvements
 
 - [#12](https://github.com/MalteHerrmann/upgrade-helper/pull/12) Add version and height links to proposal.
 - [#11](https://github.com/MalteHerrmann/upgrade-helper/pull/11) Use octocrab to access GitHub data.
 - [#10](https://github.com/MalteHerrmann/upgrade-helper/pull/10) Estimate the upgrade height based on last 50.000 blocks.
 - [#9](https://github.com/MalteHerrmann/upgrade-helper/pull/9) Add logic for selection and calculation of upgrade date and time.
-- [#5](https://github.com/MalteHerrmann/upgrade-helper/pull/5) Prepare proposal contents depending on input.
 - [#4](https://github.com/MalteHerrmann/upgrade-helper/pull/4) Check if release already exists on GitHub.
-- [#3](https://github.com/MalteHerrmann/upgrade-helper/pull/3) Add basic CLI structure.